requirements.txt
streamlit
google-genai
streamlit-mic-recorder
gtts
pandas
